Tests/LibWeb/Text/input/async-scrolling/nested-navigable-removed-during-wheel.html

<!DOCTYPE html>
<script src="../include.js"></script>
<style>
    body {
        margin: 0;
    }

    iframe {
        display: block;
        width: 160px;
        height: 120px;
        border: 0;
    }
</style>
<iframe id="iframe" srcdoc="<!DOCTYPE html><body style='height: 1000px; margin: 0'>iframe</body>"></iframe>
<script>
    promiseTest(async () => {
        const iframe = document.getElementById("iframe");
        await new Promise(resolve => iframe.addEventListener("load", resolve, { once: true }));

        let wheelEventCount = 0;
        let wheelEventWasCancelable = null;
        iframe.contentWindow.addEventListener("wheel", event => {
            wheelEventCount++;
            wheelEventWasCancelable = event.cancelable;
            iframe.remove();
        }, { passive: true });

        await animationFrame();
        await animationFrame();

        const result = await Promise.race([
            internals.wheel(50, 50, 0, 100).then(() => "resolved"),
            timeout(500).then(() => "timed out"),
        ]);

        println(`wheel promise: ${result}`);
        println(`wheel event count: ${wheelEventCount}`);
        println(`wheel event cancelable: ${wheelEventWasCancelable}`);
        println(`frame removed: ${iframe.parentNode === null}`);
    });
</script>
