Tests/LibWeb/Text/expected/async-scrolling/nested-navigable-removed-during-wheel.txt

wheel promise: resolved
wheel event count: 1
wheel event cancelable: false
frame removed: true
